refactor(select-utils): migrate to typescript
