+++ packages/components/inputs/select-utils/src/clear-indicator/index.ts
