+++ packages/components/inputs/select-utils/src/custom-styled-select-options/constants.ts
@@ -4,4 +4,4 @@ export const SELECT_DROPDOWN_OPTION_TYPES = {
   SINGLE_PROPERTY: 'single-property',
   DOUBLE_PROPERTY: 'double-property',
   MULTIPLE_PROPERTIES: 'multiple-properties',
-};
+} as const;
