+++ packages/components/inputs/select-utils/src/custom-styled-select-options/custom-styled-select-options.tsx
@@ -1,19 +1,33 @@
-import PropTypes from 'prop-types';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
-import { components } from 'react-select';
+import { components, OptionProps } from 'react-select';
 import { NO_VALUE_FALLBACK, SELECT_DROPDOWN_OPTION_TYPES } from './constants';
 
-export const MultiplePropertiesSelectInputOption = (props) => {
+type TData = {
+  label: string;
+  key: string;
+  id: string;
+};
+
+type SelectDropdownOptionTypesKeys = keyof typeof SELECT_DROPDOWN_OPTION_TYPES;
+type TProps = {
+  data?: TData;
+  noValueFallback?: string;
+  optionType?: typeof SELECT_DROPDOWN_OPTION_TYPES[SelectDropdownOptionTypesKeys];
+} & OptionProps;
+
+export const MultiplePropertiesSelectInputOption = (props: TProps) => {
   const { data } = props;
   const noValueFallback = props.noValueFallback || NO_VALUE_FALLBACK;
 
   return (
     <components.Option {...props}>
       <Spacings.Stack scale="xs">
-        <Text.Detail isBold={true}>{data.label || noValueFallback}</Text.Detail>
-        <Text.Detail>Key: {data.key || noValueFallback}</Text.Detail>
-        <Text.Detail>ID: {data.id || noValueFallback}</Text.Detail>
+        <Text.Detail isBold={true}>
+          {data?.label || noValueFallback}
+        </Text.Detail>
+        <Text.Detail>Key: {data?.key || noValueFallback}</Text.Detail>
+        <Text.Detail>ID: {data?.id || noValueFallback}</Text.Detail>
       </Spacings.Stack>
     </components.Option>
   );
@@ -21,38 +35,31 @@ export const MultiplePropertiesSelectInputOption = (props) => {
 
 MultiplePropertiesSelectInputOption.displayName =
   'MultiplePropertiesSelectInputOption';
-MultiplePropertiesSelectInputOption.propTypes = {
-  data: PropTypes.shape({
-    label: PropTypes.string,
-    key: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-    id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  }),
-  noValueFallback: PropTypes.string,
-};
 
-export const DoublePropertySelectInputOption = (props) => {
+export const DoublePropertySelectInputOption = (props: TProps) => {
   const { data } = props;
   const noValueFallback = props.noValueFallback || NO_VALUE_FALLBACK;
   return (
     <components.Option {...props}>
       <Spacings.Stack scale="xs">
-        <Text.Detail isBold={true}>{data.label || noValueFallback}</Text.Detail>
-        <Text.Detail>Key: {data.key || noValueFallback}</Text.Detail>
+        <Text.Detail isBold={true}>
+          {data?.label || noValueFallback}
+        </Text.Detail>
+        <Text.Detail>Key: {data?.key || noValueFallback}</Text.Detail>
       </Spacings.Stack>
     </components.Option>
   );
 };
 
 DoublePropertySelectInputOption.displayName = 'DoublePropertySelectInputOption';
-DoublePropertySelectInputOption.propTypes = {
-  data: PropTypes.shape({
-    label: PropTypes.string,
-    key: PropTypes.string,
-  }),
-  noValueFallback: PropTypes.string,
-};
 
-export const CustomSelectInputOption = (props) => {
+type TCustomSelectInputOptionProps = {
+  optionInnerProps: TProps;
+} & TProps;
+
+export const CustomSelectInputOption = (
+  props: TCustomSelectInputOptionProps
+) => {
   const noValueFallback = props.noValueFallback || NO_VALUE_FALLBACK;
 
   switch (props.optionType) {
@@ -74,7 +81,7 @@ export const CustomSelectInputOption = (props) => {
       return (
         <components.Option {...props.optionInnerProps}>
           <Text.Detail>
-            {props.optionInnerProps.data.label || noValueFallback}
+            {props.optionInnerProps.data?.label || noValueFallback}
           </Text.Detail>
         </components.Option>
       );
@@ -82,14 +89,3 @@ export const CustomSelectInputOption = (props) => {
 };
 
 CustomSelectInputOption.displayName = 'CustomSelectInputOption';
-CustomSelectInputOption.propTypes = {
-  optionType: PropTypes.oneOf(Object.values(SELECT_DROPDOWN_OPTION_TYPES)),
-  noValueFallback: PropTypes.string,
-  optionInnerProps: PropTypes.shape({
-    data: PropTypes.shape({
-      label: PropTypes.string,
-      key: PropTypes.string,
-      id: PropTypes.string,
-    }),
-  }).isRequired,
-};
