+++ packages/components/inputs/select-utils/src/dropdown-indicator/index.ts
