+++ packages/components/inputs/select-utils/src/search-icon-dropdown-indicator/index.ts
